+++ app/models/conference.rb
@@ -54,7 +54,7 @@ class Conference < ActiveRecord::Base
   end
 
   def after_destroy
-    joomla_article.destroy
+    joomla_article.destroy if joomla_article
   end
 
   def hosting?
