+++ app/models/session.rb
@@ -45,7 +45,7 @@ class Session < ActiveRecord::Base
   end
 
   def after_destroy
-    joomla_article.destroy
+    joomla_article.destroy if joomla_article
   end
 
   def overlaps? rhs
